refactor: split settings navigation into per-component files
